Refactor and a corner case solving
refactored auth classes and solved a case when category will be null
while fetching
