+++ src/main/java/com/example/todos/service/AuthenticationService.java
@@ -7,15 +7,15 @@ import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import com.example.todos.auth.dto.AuthenticationErrorDTO;
-import com.example.todos.auth.dto.AuthenticationRequest;
-import com.example.todos.auth.dto.AuthenticationResponse;
-import com.example.todos.auth.dto.AuthenticationTokenDTO;
-import com.example.todos.auth.dto.RegisterRequest;
 import com.example.todos.config.JwtService;
 import com.example.todos.enums.Role;
 import com.example.todos.model.TodoDAO;
 import com.example.todos.model.User;
+import com.example.todos.model.dto.AuthenticationErrorDTO;
+import com.example.todos.model.dto.AuthenticationRequest;
+import com.example.todos.model.dto.AuthenticationResponse;
+import com.example.todos.model.dto.AuthenticationTokenDTO;
+import com.example.todos.model.dto.RegisterRequest;
 import com.example.todos.model.dto.TodoResponse;
 import com.example.todos.model.dto.UserDTO;
 import com.example.todos.repository.UserRepository;
